requirements.txt
absl-py==2.2.1
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==10.4.0
protobuf==4.25.6
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.5.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.30.6
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
